Calculate TLE height systematics
This notebook tried to calculate the systematics on sat. height given
a set of TLEs which might not come from the same source.

I don't know if this assumption is true and this is probably not reliable.
My attempts to talk to experts about this was not successful.  This was
not needed in the end as I was able to the the x,y,z data without TLEs.